2841: Resolve StringUtils header (#2863)
* 2841: refactor and split up string utils

* 2841: make KDL lib static, allows to hide STL includes from string utilities

* 2841: resolving StringUtils wip

* 2841: remove StringUtils header

* 2841: fix build errors

# Conflicts:
#	common/src/IO/Path.h

* 2841: fix more build errors

* 2841: enable implicit conversion warnings for clang and gcc, fix a few build problems

* 2841: add missing include

* 2841: fix implicit conversion warnings from GCC

# Conflicts:
#	common/test/src/Model/BrushTest.cpp

* 2841: make kdl header-only again

* 2841: add missing tests for kdl

* 2841: revert change to str_is_numeric

+++ common/src/Renderer/RenderUtils.cpp
@@ -89,7 +89,7 @@ namespace TrenchBroom {
 
             std::vector<vm::vec2f> vertices(segments + 1);
 
-            const float d = angleLength / segments;
+            const float d = angleLength / static_cast<float>(segments);
             float a = startAngle;
             for (size_t i = 0; i <= segments; ++i) {
                 vertices[i][0] = radius * std::sin(a);
@@ -121,7 +121,7 @@ namespace TrenchBroom {
                     break;
             }
 
-            const float d = angleLength / segments;
+            const float d = angleLength / static_cast<float>(segments);
             float a = startAngle;
             for (size_t i = 0; i <= segments; ++i) {
                 vertices[i][x] = radius * std::cos(a);
@@ -178,7 +178,7 @@ namespace TrenchBroom {
             vertices.resize(roundedRect2DVertexCount(cornerSegments));
             size_t vertexIndex = 0;
 
-            const float angle = vm::Cf::half_pi() / cornerSegments;
+            const float angle = vm::Cf::half_pi() / static_cast<float>(cornerSegments);
             vm::vec2f center(0.0f, 0.0f);
             vm::vec2f translation;
